Add option to increase the graph radius when plotting central species
An optional radius argument can now be used to increase the graph
radius of the flux diagram when using central species. In addition, it
is now also possible to superimpose central species onto the entire
flux diagram to ensure that they appear in it. This is especially
useful for systems in which species are produced via low-flux
channels, yet where they accumulate to significant concentrations.