Reverse reactant order when generating reaction with swapped reactants
Previously, when checking reaction generation from swapped reactions
in bimolecular families, only the reactant template order was swapped,
but not the order of reactants. For families with the same template
for both reactants, this means that not all products can be generated
if they depend on the order of reactants. The dependence on the
reactant order originates from the relabeling of atoms that happens
when applying the recipe.